Enabling MPFR on Windows
git-svn-id: https://svn.cct.lsu.edu/repos/projects/parallex/trunk/hpx@2840 5079c8e4-9419-0410-93bb-9e5ee278c886

+++ examples/had_amr/amr_c/stencil.cpp
@@ -301,7 +301,7 @@ namespace hpx { namespace components { namespace amr
         //threads::thread_id_type id = self.get_thread_id();
         //threads::set_thread_description(id,description);
 
-        if (val[compute_index]->timestep_ < numsteps_) {
+        if (val[compute_index]->timestep_ < (int)numsteps_) {
 
             // copy over critical info
